--- policy-tool-frontend-v2/src/components/FormSection/index.tsx
@@ -1,1 +1,0 @@
-export * from './FormSection'
