--- policy-tool-frontend-v2/src/components/InputWrapper/index.ts
@@ -1,1 +1,0 @@
-export * from '../InputWrapper'
