--- policy-tool-frontend-v2/src/components/Layout/index.ts
@@ -1,1 +1,0 @@
-export * from './Layout'
