--- policy-tool-frontend-v2/src/components/LoadingWrapper/index.ts
@@ -1,1 +1,0 @@
-export * from './LoadingWrapper'
