--- policy-tool-frontend-v2/src/components/MenuButton/index.ts
@@ -1,1 +1,0 @@
-export * from './MenuButton'
