+++ policy-tool-frontend-v2/src/components/Selector.tsx
