--- policy-tool-frontend-v2/src/components/Selector/index.ts
@@ -1,1 +1,0 @@
-export * from './Selector'
